requirements.txt
Jinja2==3.1.2
markdown==3.4.3
python-frontmatter==1.0.0